atlahs: harden RNIC driver bootstrap

+++ htsim/sim/lgs/Parser.hpp
@@ -859,12 +859,12 @@ class Parser {
 		}
 	}
 
-	uint8_t GetNumCPU() {
-		return max_cpu+1;
+	uint32_t GetNumCPU() {
+		return static_cast<uint32_t>(max_cpu) + 1;
 	}
 
-	uint8_t GetNumNIC() {
-		return max_nic+1;
+	uint32_t GetNumNIC() {
+		return static_cast<uint32_t>(max_nic) + 1;
 	}
 
 
@@ -875,4 +875,3 @@ class Parser {
 	}
 	
 };
-
